Add files via upload
For the meeting on April 16th, 2020, which will likely be a more efficient occasion to give a full summary.

A script that iterates through the .csv files associated with the PEPI/LVR database, using regex to identify target information. Afterwards, the information associated with each section is split into key-value pairs that are placed into various dictionaries, where data can be retrieved and further analysed with trivial difficulty.

Currently, only to showcase that the program can perform arbitrary file I/O, the output is a very general text file summarizing a few data points gathered from each sheet. This is totally disconnected from the entirety of the information the script as a whole collects. With the current setup, it is very straightforward to loop through the dictionaries associated with each sheet to analyse data and generate whatever final product is desired.